+++ src/app/components/pages/Account/tabs/AccountStorageTab.tsx
@@ -27,7 +27,10 @@ export const AccountStorageTab = memo((props: Props) => {
 
     const { launcherThunks } = useThunks();
 
-    const { result: s3MustacheParams } = useAsync(launcherThunks.getS3MustacheParams, []);
+    const { result: s3MustacheParams } = useAsync(
+        launcherThunks.getS3MustacheParamsForProjectBucket,
+        [],
+    );
 
     const onRequestCopyFactory = useCallbackFactory(([textToCopy]: [string]) =>
         copyToClipboard(textToCopy),
@@ -148,7 +151,7 @@ export declare namespace AccountStorageTab {
     };
 }
 
-const useStyles = makeStyles({ "label": { AccountStorageTab } })(theme => ({
+const useStyles = makeStyles({ "name": { AccountStorageTab } })(theme => ({
     "divider": {
         ...theme.spacing.topBottom("margin", 4),
     },